debugged: works perfectly on ubuntu
