+++ src/get_source_from_cvepatch.py
@@ -1,8 +1,4 @@
 #!/usr/bin/env python
-"""
-This module retrieves the vulnerable function from cvepatch (diff)
-Author: Seulbae Kim (seulbae@korea.ac.kr)
-"""
 
 import os
 import sys
